requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
discord.py==2.4.0
exceptiongroup==1.2.1
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3
tomli==2.0.1
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
